LifeCache/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
streamlit==1.39.0
vaderSentiment==3.3.2
sumy==0.11.0
scikit-learn==1.5.2
speechrecognition==3.10.4
pocketsphinx==5.0.4
fpdf2==2.7.9
matplotlib==3.9.2
apscheduler==3.10.4
python-multipart==0.0.12
